requirements.txt
flask
sty
numpy
pillow
requests
waitress
flask-limiter
